Further major refactoring. Starting to set up VCF only part
